backend/requirements.txt
OPENAI_API_KEY= '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'

aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.13.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cssselect==1.2.0
distro==1.9.0
edge-tts==7.0.0
exceptiongroup==1.2.2
fastapi==0.115.8
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.17.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jieba3k==0.35.1
jiter==0.8.2
joblib==1.4.2
lxml==5.3.0
multidict==6.1.0
newspaper3k==0.2.8
nltk==3.9.1
openai==1.61.0
pillow==11.1.0
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
srt==3.5.3
starlette==0.45.3
tabulate==0.9.0
tinysegmenter==0.3
tldextract==5.1.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
